\name{causalMediation-package}
\alias{causalMediation-package}
\docType{package}
\title{
Causal Mediation Analysis.
}
\description{
Implementation of Causal Mediation Analysis as published by Valeri and VanderWeele (2013, doi: 10.1037/a0031034).
This package does not implement the delta method for estimation of the standard errors but uses bootstrapping in stead.
}
\details{
\tabular{ll}{
Package: \tab causalMediation\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-09-07\cr
License: \tab GPL-2\cr
}
}
\author{
Egge van der Poel

Maintainer: Egge van der Poel <e.vanderpoel@erasmusmc.nl>
}
\references{
Valeri and VanderWeele (2013, doi: 10.1037/a0031034)
}
\keyword{ causalMediation }
\seealso{
None
}
